Extract remaining inventory mutation modules
